refactor(agreement): 약관 동의 공통 컴포넌트로 분리
