+++ src/app/(route)/auth/kakao/callback/_components/index.ts
@@ -1,2 +1,2 @@
 export { default as KakaoContainer } from "./KakaoContainer/KakaoContainer";
-export { default as TermAgree } from "./TermAgree/TermAgree";
+export { default as KakaoLoading } from "./KakaoLoading/KakaoLoading";
