+++ src/app/(route)/sign-up/_components/index.ts
@@ -1,3 +1,2 @@
-export { default as AllAgree } from "./AllAgree/AllAgree";
 export { default as SignUpField } from "./SignUpField/SignUpField";
 export { default as SignUpContainer } from "./SignUpContainer/SignUpContainer";
