+++ src/components/domain/TermsAgreement/_constants/TERMS_CONFIG.ts
