+++ src/components/domain/index.ts
@@ -25,6 +25,7 @@ export { default as ProfileEditSection } from "./ProfileForm/ProfileEditSection"
 export { default as BlockUserModal } from "./PostReportBlockActions/UserBlockModal/UserBlockModal";
 export { default as AddToHomeScreenPWA } from "./AddToHomeScreenPWA/AddToHomeScreenPWA";
 export { default as ContentShareModal } from "./ContentShareModal/ContentShareModal";
+export { default as TermsAgreement } from "./TermsAgreement/TermsAgreement";
 
 // TODO: 베타 테스트 종료 후 제거
 export { default as BetaTestMypageBanner } from "./BetaTest/BetaTestMypageBanner/BetaTestMypageBanner";
